fixed minor bug

+++ Pokemon.java
@@ -1,6 +1,6 @@
 public abstract class Pokemon{
     protected int _hp, _speed, _spatk, _atk, _spdef, _def;
-    protected float _currhp;
+    protected int _currhp;
     protected String _type1, _type2, _name;  //type2 can be Null
     protected String _statusEffect;
     protected Object[] _statMods;
